Home

ParseTrees.jl

ParseTrees.parse_tree

ParseTrees.separate_clauses

ParseTrees.parse_tree — Method.

parse_tree(file, format = "enhanced-plus-plus-dependencies")

Process the results of the coreNLP parse annotator from an XML file.

For each sentence, will return meta data about each word as well as a tree of how the words are connected.

You can run coreNLP through the command line. A minimal command is java -cp "*" edu.stanford.nlp.pipeline.StanfordCoreNLP -annotators tokenize,ssplit,parse -fileList FILE_LIST from the folder where you unzipped coreNLP. The resulting files will be in the format FILE.xml, each of which you can parse with parse_tree.

julia> using ParseTrees

julia> result = parse_tree("hammurabi.txt.xml");

julia> result[31].meta
17-element Array{NamedTuple{(:relationship, :text),Tuple{String,String}},1}:
 (relationship = "mark", text = "If")
 (relationship = "det", text = "any")
 (relationship = "nsubj", text = "one")
 (relationship = "advcl:if", text = "steal")
 (relationship = "det", text = "the")
 (relationship = "amod", text = "minor")
 (relationship = "dobj", text = "son")
 (relationship = "case", text = "of")
 (relationship = "nmod:of", text = "another")
 (relationship = "punct", text = ",")
 (relationship = "nsubjpass", text = "he")
 (relationship = "aux", text = "shall")
 (relationship = "auxpass", text = "be")
 (relationship = "root", text = "put")
 (relationship = "case", text = "to")
 (relationship = "nmod:to", text = "death")
 (relationship = "punct", text = ".")

julia> result[31].tree
{16, 15} directed simple Int64 graph

source

ParseTrees.separate_clauses — Method.

separate_clauses(sentence, clause_types)

Pull out root level clauses from a parsed sentence that are one of the clause_types.

julia> using ParseTrees

julia> sentence = parse_tree("hammurabi.txt.xml")[31];

julia> result = separate_clauses(sentence, ("advcl:if", "nsubjpass", "aux"));

julia> result.clauses[1]
(clause_type = "advcl:if", text = "If any one steal the minor son of another")

julia> result.clauses[2]
(clause_type = "nsubjpass", text = "he")

julia> result.clauses[3]
(clause_type = "aux", text = "shall")

julia> result.rest
", be put to death"

source