Home

ParseTrees.jl

ParseTrees.annotate

ParseTrees.clauses

ParseTrees.dependencies

ParseTrees.sentences

ParseTrees.annotate — Function.

annotate(file, annotators = ("tokenize", "ssplit", "parse"), options = ``)

Return a command line call to coreNLP to be run.

Must be run with the working directory set to the location where you have unzipped t coreNLP. Will take a few minutes to complete.

julia> using ParseTrees

julia> annotate("hammurabi.txt")
`java -cp '*' edu.stanford.nlp.pipeline.StanfordCoreNLP -annotators tokenize,ssplit,parse -file hammurabi.txt`

source

ParseTrees.clauses — Method.

clauses(tree, meta, clause_types)

Pull out root level clauses that are of clause_types.

Input the the meta data and dependency tree that result from dependencies.

julia> using ParseTrees

julia> depencies_31 = collect(sentences("hammurabi.txt.xml"))[31] |> dependencies;

julia> clauses_31 = clauses(depencies_31.tree, depencies_31.meta, ("advcl", "nsubjpass", "aux"));

julia> clauses_31.clauses[1]
(clause_type = "advcl", text = "If any one steal the minor son of another")

julia> clauses_31.clauses[2]
(clause_type = "nsubjpass", text = "he")

julia> clauses_31.clauses[3]
(clause_type = "aux", text = "shall")

julia> clauses_31.rest
", be put to death"

source

ParseTrees.dependencies — Function.

dependencies(sentence, format = "basic-dependencies")

Return meta data about each token as well as a tree of how the tokens are connected.

Input one of the sentences of a coreNLP run containing at least the tokenize, ssplit, and parse annotators.

julia> using ParseTrees

julia> depencies_31 = collect(sentences("hammurabi.txt.xml"))[31] |> dependencies;

julia> depencies_31.tree
{16, 15} directed simple Int64 graph

julia> depencies_31.meta
17-element Array{NamedTuple{(:relationship, :text),Tuple{String,String}},1}:
 (relationship = "mark", text = "If")
 (relationship = "det", text = "any")
 (relationship = "nsubj", text = "one")
 (relationship = "advcl", text = "steal")
 (relationship = "det", text = "the")
 (relationship = "amod", text = "minor")
 (relationship = "dobj", text = "son")
 (relationship = "case", text = "of")
 (relationship = "nmod", text = "another")
 (relationship = "punct", text = ",")
 (relationship = "nsubjpass", text = "he")
 (relationship = "aux", text = "shall")
 (relationship = "auxpass", text = "be")
 (relationship = "root", text = "put")
 (relationship = "case", text = "to")
 (relationship = "nmod", text = "death")
 (relationship = "punct", text = ".")

source

ParseTrees.sentences — Method.

sentences(file)

Return an iterator over sentences.

Input the XML results a run of coreNLP including at least the tokenize and ssplit annotators.

julia> using ParseTrees

julia> result = sentences("hammurabi.txt.xml");

julia> typeof(first(result))
LightXML.XMLElement

source