ParseTrees.jl

ParseTrees.jl

ParseTrees.institutional_grammar

ParseTrees.flat

ParseTrees.rules

ParseTrees.sentences

ParseTrees.institutional_grammar — Constant.

institutional_grammar

A clause dict based on Elinor Ostrom's institutional grammar.

source

ParseTrees.flat — Method.

flat(sentence)

Just add water

flat(sentence, id)

Just the parts of the sentence connected to id.

source

ParseTrees.rules — Function.

rules(file, clause_dict, rest = :rest)

Split the sentences of a file into groups of clauses based on clause_dict, starting at the root. Clause dict should be a map from Universal Dependencies (v1) to clause categories. There are three reserved clause categories, :recur, :remove, and rest.

julia> using ParseTrees

julia> result = rules("hammurabi.txt.xml", institutional_grammar, :aIm);

julia> result[237]
5-element Array{Pair{Symbol,String},1}:
 :Condition => "If any one hire a cart alone"
 :Attribute => "he"
   :Deontic => "shall"
    :Object => "forty ka of corn per day"
       :aIm => "pay"

source

ParseTrees.sentences — Method.

sentences(file)

Return an iterator over sentences of an XML coreNLP result.

You can run coreNLP through the command-line. This package only requires three annontators: tokenize, ssplit, and parse.

julia> using ParseTrees

julia> result = sentences("hammurabi.txt.xml");

julia> result[2]
{29, 28} directed Int64 metagraph with Float64 weights defined by :weight (default weight 1.0)

julia> flat(result[2])
"If any one ensnare another , putting a ban upon him , but he can not prove it , then he that ensnared him shall be put to death"

julia> flat(result[2], 4)
"If any one ensnare another , putting a ban upon him"

source