ParseTrees.jl

ParseTrees.jl

ParseTrees.deontic

ParseTrees.dictionary

ParseTrees.flat

ParseTrees.sentences

ParseTrees.split_rules

ParseTrees.deontic — Constant.

const deontic

Default pattern to recognize deontics.

source

ParseTrees.dictionary — Constant.

dictionary

A dictionary based on Elinor Ostrom's institutional grammar.

source

ParseTrees.flat — Method.

flat(sentence)

The whole sentence

flat(sentence, root)

Just the parts of the sentence connected to root.

source

ParseTrees.sentences — Method.

sentences(file)

Return an iterator over sentences of an coreNLP rules XML file.

You can run coreNLP through the command-line. This package only uses the parse annotator.

julia> using ParseTrees

julia> sentence = sentences("hammurabi.txt.xml")[2]
{29, 28} directed Int64 metagraph with Float64 weights defined by :weight (default weight 1.0)

julia> flat(sentence)
"If any one ensnare another , putting a ban upon him , but he can not prove it , then he that ensnared him shall be put to death"

julia> flat(sentence, 4)
"If any one ensnare another , putting a ban upon him"

source

ParseTrees.split_rules — Method.

split_rules(file; dictionary = dictionary, deontic = deontic)

Split file into rules. Split each rule into components based on dictionary, starting at the root. The dictionary should be a map from Universal Dependencies (v1) to clause components. :deontics must additionally match the deontic pattern. Rules must contain a deontic. :recur will look for a rule inside the rule, :remove will remove the clause, and :root will include the root.

julia> using ParseTrees

julia> parsed = split_rules("hammurabi.txt.xml");

julia> parsed[25]
6-element Array{Pair{Symbol,String},1}:
 :condition => "If the slave will not give the name of the master"
 :attribute => "the finder"
   :deontic => "shall"
    :object => "him"
 :condition => "to the palace"
      :root => "bring"

source