ParseTrees.jl

ParseTrees.jl

ParseTrees.clauses

ParseTrees.dependencies

ParseTrees.rules

ParseTrees.sentences

ParseTrees.clauses — Method.

clauses(dependencies_result, clause_types)

Pull out root level clauses that are of clause_types from the result of dependencies.

source

ParseTrees.dependencies — Method.

dependencies(sentence)

Return meta-data about each token as well as a tree of how the tokens are connected.

source

ParseTrees.rules — Method.

rules(document)

Find the institutional grammar components in a coreNLP result.

julia> using ParseTrees

julia> result = rules("hammurabi.txt.xml");

julia> length(result)
234

julia> result[2]
5-element Array{Pair{Symbol,String},1}:
 :Attribute => "his accuser"
   :Deontic => "shall"
       :aIm => "take"
 :Condition => "If any one bring an accusation against a man , and the accused go to the river and leap into the river , if he sink in the river"
    :Object => "possession of his house"

source

ParseTrees.sentences — Method.

sentences(file)

Return an iterator over sentences of an XML coreNLP result.

You can run coreNLP through the command-line. This package only requires three annontators: tokenize, ssplit, and parse.

source